Dont panic on template parse failure and change delims
the default delimitators were changed to be [[ ]] but unfortunately
there is a high chance of hitting those delimitators in a cloud config
file that can be using commands in its steps.

This patch changes those delimitators to [[[ ]]] which is much more
difficult to hit.

It also reworks the render method to be more careful to avoid raising a
panic and leaving the user with no clue on what happened. Now it
properly checks for errors when parsing the data and returns that.

It alos logs the error properly to the logger and to debug with much
more data so its easier to troubleshoot

Signed-off-by: Itxaka <itxaka@kairos.io>