Simplify function and improve log
The pixiecore spec allows more than one initrd for some reason, but we
dont use that at all. Simplify that by just passing the plain initrd and
using it directly.

Also drop the bootmsg, its not necessary and it would need a very VERY
slow netwok to show up.

Also make the log the same as we use in aurora, splitting it to debug
and info so the logging is nicer and falls in line with the previous log

Signed-off-by: Itxaka <itxaka@kairos.io>