Remove not necessary flag
it seems that it only fails when the user building is not the same as
the owner of the .git directory:

https://github.com/golang/go/issues/53532

I only saw it when running:

```
docker run --rm -it -v $PWD:/work --workdir /work golang go build .
```

With `docker build` it just works (maybe that command handles the users
differently?). In any case, it doesn't affect CI or anything else so
I'll remove it.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>