Do not change name to extend-cmdline output files
For some weird reason we were changing the name for the extended cmdline
efi output file, which is worn. It wont collide with anything as we are
just generating a single file, so there is no need to change its name.

This generate weird names like norole_EXTENDEDCMDLINE.efi and conf files
which then did not pass the upgrade requirements (which checks for
norole.efi)

Signed-off-by: Itxaka <itxaka@kairos.io>