Switch to Fedora for the auroraboot image (#125)
* Switch to Fedora for the auroraboot image

because opensuse/leap doesn't have latest systemd and older systemd-repart
doesn't support the flags we are passing (e.g.  --make-ddi).

Our other option would be to switch to Tumbleweed but that might be too
unstable for us to use for Auroraboot.

This is reviving the work here:
https://github.com/kairos-io/osbuilder/pull/128

later reverted here:
https://github.com/kairos-io/osbuilder/commit/7813c2eec45a1db8ff9b1283ed2c9fa73e4cf878

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Add binutils (for objcopy)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Add qemu-img and qemu-system-x86_64

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

---------

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>