Fix different behaviour between usb and cd in efi (#135)
* Use the efi image directly and set efi_boot_part

this supposedly increases compatibility when booting from usb

Mark the current El Torito boot image (see options -b and -e) in an actually invalid GPT as partition of type Basic Data. This works only with -isohybrid-mbr and has the same impact on the system area as -efi-boot-part. It cannot be combined with -efi-boot-part or -hfsplus.
The first three boot images which are marked by GPT will also show up as partition entries in MBR. The MBR partition of type 0xEF is what actually is used by EFI firmware for booting from USB stick. The MBR partition for PC-BIOS gets type 0x00 rather than 0x17 in this case. Often the further MBR entries are the ones which actually get used by EFI.

Signed-off-by: Itxaka <itxaka@kairos.io>

* Make it work for livecd and usb

Signed-off-by: Itxaka <itxaka@kairos.io>

* Review comments

Signed-off-by: Itxaka <itxaka@kairos.io>

* Apply suggestions from code review

Co-authored-by: Mauro Morales <mauro.morales@spectrocloud.com>

* Move dir creation to function

Signed-off-by: Itxaka <itxaka@kairos.io>

---------

Signed-off-by: Itxaka <itxaka@kairos.io>
Co-authored-by: Mauro Morales <mauro.morales@spectrocloud.com>