Cleanup temp dirs properly (#145)
We create a couple of temp dirs when starting aurora, it makes sense to
remove them when we are done as they are used for several transient
things.

Plus it has been reported that leaving things around can break building
new artifacts, as we leave kernels around and they are picked up instead
of the current ones, so its jsut good behaviour to clean up.

this cleans up before creating the temp dirs (fixed names) and when
finished

Signed-off-by: Itxaka <itxaka@kairos.io>