Add tracking of time it takes each step to execute
Useful to track how much do we take normally in the steps and improve on
places that there may be a bottle neck or help us run things in parallel
if possible.

Its set at trace level, which IIRC its not possible to set yet, so it
needs a way of enabling trace level, maybe by env var or something?

Signed-off-by: Itxaka <itxaka@kairos.io>