Don't shell out to the same binary (#258)
calling `auroraboot` from within the `auroraboot` go binary is weird and
makes development harder (since a simple `go run .` doesn't cut it, you
also need to build the `auroraboot` binary and put it in the PATH).
We probalby did it because of the packages not being very well
structured, easily leading to circular imports.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>