Add "worker" concept
We will allow the user to spin up a "local" worker next to the server
(for completely local builds like those we had up to now). We will also
allow the user to spin up remote workers too (which we can use to deploy
auroraboot in Kubernetes or otherwise, horizontally scaled).

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>