Fix logs not appearing in the websocket stream
Some of these "fixes" might not be always needed (e.g. wait before we
close the socket) but better safe than sorry. The main issue was that
some auroraboot libraries were creating their own loggers instead of
using the internal logger. Thus the logs were written in the stdout but
not sent to the writter we specified (websockets).

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>