3308 factory api (#274)
* Introduce API endpoints for job building

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Add "worker" concept

We will allow the user to spin up a "local" worker next to the server
(for completely local builds like those we had up to now). We will also
allow the user to spin up remote workers too (which we can use to deploy
auroraboot in Kubernetes or otherwise, horizontally scaled).

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Implement logs endpoints

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Make the worker send logs to the websockets endpoint

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Create (optional) worker along with the server, for local builds

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* [WIP] Migrate to a filesystem based storage for builds

because:

- we want then the server dies (e.g. Pod restarts) to continue where we
  left-off (persistence of jobs)
- we don't want to introduce a database
- we want to also store the artifacts. Those will be uploaded using a
  new endpoint, from the worker, to the server.

Missing:

- locking mechanism to make the updates of the job metadata atomic
- upload endpoint for the artifacts

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Make binding a job an atomic operation

and prepare the code to be extracted to an interface in case we want to
use a database in production.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Move the build logic to the worker

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Fix download links

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Run all non-mutually-exclusive ifs

because a message may include more than one of the matching strings
(hard to happen but still)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Fix logs not appearing in the websocket stream

Some of these "fixes" might not be always needed (e.g. wait before we
close the socket) but better safe than sorry. The main issue was that
some auroraboot libraries were creating their own loggers instead of
using the internal logger. Thus the logs were written in the stdout but
not sent to the writter we specified (websockets).

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Send only plain text from the worker

to avoid all the complexity on the javascript side.
Especially since there is writting to a file in the middle, we can't be
sure we got the full json on the js side.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Fix new lines missing from logs

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Sanitize user provided input at the lowest level

If the user tries to traverse the filesystem with a maliciously crafter
job id, our code will return an error. We only accept valid UUIDs for
job ids.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Fix worker tests

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Use a message that matches any docker version

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Revert un-intended changes

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Allow empty lines and sanitize input by populating textContent

so that there is no way to inject html through the logs

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Move all initilization args to the AppConfig (PR comment)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Change the mandatory fields

In latest kairos-init, "registry" is also mandatory but the Web UI
doesn't have such a field either. I'm also considering keeping it
optional this way:
https://github.com/kairos-io/AuroraBoot/pull/274#discussion_r2088864835

For now, since the references kairos-init in the Dockerfile doesn't
require the "registry", I'll leave it out.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Use the "address" Flag properly when spinning up the worker

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Add some logs to the worker (to know what's going on)

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

* Use a multi-writer everywhere

so that all logs appear both in stdout and also over the wire
(websockets).

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>

---------

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>