Change base_image to a new type of field
It's now just a text box with helper buttons to fill-in predefined
values. The user can either edit those or replace completely with their
own base image. It simplifies the logic by a lot because, after all,
it's only one value we need, so there should be only one input field for
it.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>